Split up Key
